+++ packages/react-strict-dom/src/native/modules/createStrictDOMTextComponent.js
@@ -7,7 +7,8 @@
  * @flow strict-local
  */
 
-import type { StrictProps } from '../../types/StrictProps';
+import type { ReactNativeProps } from '../../types/renderer.native';
+import type { StrictProps as StrictPropsOriginal } from '../../types/StrictProps';
 
 import * as React from 'react';
 import { Animated, Platform, Text } from 'react-native';
@@ -19,6 +20,11 @@ import { useNativeProps } from './useNativeProps';
 import { useStrictDOMElement } from './useStrictDOMElement';
 import * as stylex from '../stylex';
 
+type StrictProps = $ReadOnly<{
+  ...StrictPropsOriginal,
+  children?: React.Node | ((ReactNativeProps) => React.Node)
+}>;
+
 function hasElementChildren(children: mixed): boolean {
   return children != null && typeof children !== 'string';
 }
@@ -26,7 +32,7 @@ function hasElementChildren(children: mixed): boolean {
 export function createStrictDOMTextComponent<T, P: StrictProps>(
   tagName: string,
   _defaultProps?: P
-): React.AbstractComponent<P, T> {
+): component(ref?: React.RefSetter<T>, ...P) {
   const component: React.AbstractComponent<P, T> = React.forwardRef(
     function (props, forwardedRef) {
       let NativeComponent = Text;
@@ -55,7 +61,7 @@ export function createStrictDOMTextComponent<T, P: StrictProps>(
       // Tag-specific props
 
       if (tagName === 'a') {
-        nativeProps.role = nativeProps.role ?? 'link';
+        nativeProps.role ??= 'link';
         if (href != null) {
           nativeProps.onPress = function (e) {
             if (__DEV__) {
@@ -99,10 +105,18 @@ export function createStrictDOMTextComponent<T, P: StrictProps>(
        * Construct tree
        */
 
-      // $FlowFixMe
-      let element: $FlowFixMe = <NativeComponent {...nativeProps} />;
+      let element: React.Node =
+        typeof props.children === 'function' ? (
+          props.children(nativeProps)
+        ) : (
+          // $FlowFixMe
+          <NativeComponent {...nativeProps} />
+        );
 
-      if (hasElementChildren(nativeProps.children)) {
+      if (
+        hasElementChildren(nativeProps.children) ||
+        typeof props.children === 'function'
+      ) {
         if (inheritableStyle != null) {
           element = (
             <ProvideInheritedStyles