+++ packages/react-strict-dom/src/types/renderer.web.js
@@ -3,8 +3,11 @@
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
+ *
+ * @flow strict
  */
 
-module.exports = {
-  presets: [require.resolve('@docusaurus/core/lib/babel/preset')]
-};
+export type ReactDOMStyleProps = $ReadOnly<{
+  className?: string,
+  style?: $ReadOnly<{ [string]: string | number }>
+}>;